requirements.txt
scipy
matplotlib
moviepy
pillow